requirements.txt
cryptography
kazoo
pyjwt
requests
